[Add] BoardDetail 페이지마다 댓글 매핑

+++ src/main/java/jungle/fairyTeller/board/dto/CommentDto.java
@@ -13,7 +13,7 @@ import java.util.Date;
 @AllArgsConstructor
 public class CommentDto {
     private Long commentId;
-    private Long boardId;
+    private Integer boardId;
     private String content;
     private String author;
     private Date createdDatetime;
